requirements.txt
argcomplete==1.11.1
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.9
importlib-metadata==1.6.0
PyYAML==5.3.1
requests==2.23.0
RPi.GPIO==0.7.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.9
xmltodict==0.12.0
yq==2.10.1
zipp==3.1.0
